refactor: move map functions to helpers
