+++ src/components/DownloadAndShareComponents/helpers/getDownloadFormErrorMessage.tsx
@@ -1,4 +1,4 @@
-import type { ErrorKeysType } from "../downloadsAndShare.types";
+import type { ErrorKeysType } from "../downloadAndShare.types";
 
 type ErrorMessagesAndOrderType = {
   order: number;
